Feat/week4/coupon (#21)
* doc: 설계서에 쿠폰 추가

* doc: 설계서에 쿠폰 추가

* feat: 쿠폰/유저쿠폰/쿠폰 내역 도메인 모델/테스트코드 추가

* feat: 유저의 보유 쿠폰 가져오기 기능 추가

* feat: 주문에 포인트 적용

* feat: 주문에 쿠폰 할인 적용

* refactor: 쿠폰 할인 정책을 UserCoupon이 주는 것에서 외부 객체로 분리

* feat: 쿠폰 동시성 문제 낙관성 락 적용

+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentInfo.java
@@ -1,9 +1,11 @@
 package com.loopers.domain.payment;
 
+import java.math.BigDecimal;
+
 public class PaymentInfo {
     public record Pay(
             Long paymentId,
-            Long paymentAmount,
+            BigDecimal paymentAmount,
             PaymentType type
     ){
         static public Pay from(Payment payment) {
